Further feedback fixes
- Create a "compatibility" package in ouroboros/chainsync, for code that can read v5 or v6 entries. (Done in part to avoid circular dependencies.)
- Small fixes found during further unit testing.
